# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "service in Go using Gin framework",
        "title": "Olt Service API",
        "contact": {
            "name": "Juan Acosta",
            "url": "https://www.linkedin.com/in/juan-m-acosta-f-54219758/",
            "email": "juan9acosta@gmail.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:7002",
    "basePath": "/",
    "paths": {
        "/cron/olt-autowrite": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to write config to olt permanently",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task olt_autowrite",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cron/olt-cleaning": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to clean old data relating olt in DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task clean_olt_data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cron/olt-getclock": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to get status of clock on all olts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task get_clock",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cron/olt-getinfo": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to get olts general info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task get_olt_info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cron/onu-cleaning": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to clean old data relating to onus in DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task clean_onu_data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cron/onu-getinfo": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to get onus general info\n-onu-status \tevery 1min\n-onu-rx \t\t\tevery 5min\n-onu-sn \t\t\tevery 30min\n-onu-name \t\tevery 30min\n-onu-ethlist\tsinc manual\n-onu-tx \t\t\tsinc manual 30min",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task get_onu_info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cron/onu-traffic": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "run cron to get onus traffic info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crons"
                ],
                "summary": "Run the task get_onu_traffic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "notice": {
                    "type": "string"
                },
                "record": {}
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}